third_party/WebKit/LayoutTests/crypto/rsa-es-generateKey-parseAlgorithm-failures.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests generateKey() with bad RSAES-PKCS1-v1_5 parameters");
jsTestIsAsync = true;

extractable = true;
keyUsages = ['encrypt', 'decrypt'];

Promise.resolve(null).then(function() {
    return crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5', modulusLength: -30}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5', modulusLength: NaN}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5'}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5', modulusLength: 10}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5', modulusLength: 10, publicExponent: 10}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5', modulusLength: 10, publicExponent: null}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);
}).then(finishJSTest, failAndFinishJSTest);

</script>

</body>
</html>
