third_party/WebKit/LayoutTests/crypto/rsa-es-generateKey-parseAlgorithm-failures-expected.txt

Tests generateKey() with bad RSAES-PKCS1-v1_5 parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


error is: SyntaxError: RsaKeyGenParams: modulusLength: Outside of numeric range
error is: SyntaxError: RsaKeyGenParams: modulusLength: Is not a number
error is: SyntaxError: RsaKeyGenParams: modulusLength: Missing required property
error is: SyntaxError: RsaKeyGenParams: publicExponent: Missing or not a Uint8Array
error is: SyntaxError: RsaKeyGenParams: publicExponent: Missing or not a Uint8Array
error is: SyntaxError: RsaKeyGenParams: publicExponent: Missing or not a Uint8Array
PASS successfullyParsed is true

TEST COMPLETE
